+++ client/django-formset/IncompleteSelect.ts
@@ -4,7 +4,7 @@ export abstract class IncompleteSelect {
 	private readonly fieldName?: string;
 	protected isIncomplete: boolean;
 	protected readonly fieldGroup: Element;
-	protected filterByValues = new Map<string, string>();
+	protected filterByValues = new Map<string, string | string[]>();
 
 	constructor(element: HTMLSelectElement) {
 		const fieldGroup = element.closest('django-field-group');
@@ -28,18 +28,27 @@ export abstract class IncompleteSelect {
 		const filters = element.getAttribute('filter-by')?.split(',') ?? [];
 		filters.forEach(filterBy => {
 			const observedElement = element.form?.elements.namedItem(filterBy);
-			if (observedElement instanceof HTMLInputElement || observedElement instanceof HTMLSelectElement) {
+			if (observedElement instanceof HTMLInputElement) {
 				this.filterByValues.set(filterBy, observedElement.value);
-				if (Array.from(this.filterByValues.values()).some(val => val)) {
-					this.reloadOptions();
-				}
 				observedElement.addEventListener('change', (event: Event) => {
 					const changedElement = event.currentTarget;
-					if (changedElement instanceof HTMLInputElement || changedElement instanceof HTMLSelectElement) {
+					if (changedElement instanceof HTMLInputElement) {
 						this.filterByValues.set(filterBy, changedElement.value);
 						this.reloadOptions();
 					}
 				});
+			} else if (observedElement instanceof HTMLSelectElement) {
+				this.filterByValues.set(filterBy, Array.from(observedElement.selectedOptions).map(o => o.value));
+				observedElement.addEventListener('change', (event: Event) => {
+					const changedElement = event.currentTarget;
+					if (changedElement instanceof HTMLSelectElement) {
+						this.filterByValues.set(filterBy, Array.from(observedElement.selectedOptions).map(o => o.value));
+						this.reloadOptions();
+					}
+				});
+			}
+			if (Array.from(this.filterByValues.values()).some(val => (val as Array<string>).some(s => s))) {
+				this.reloadOptions();
 			}
 		});
 	}
@@ -55,22 +64,27 @@ export abstract class IncompleteSelect {
 		this.fieldGroup.classList.add('dj-touched');
 	}
 
-	protected buildFetchQuery(searchStr?: string) {
-		const parts = [];
+	protected buildFetchQuery(offset: number, searchStr?: string) : URLSearchParams {
+		const query = new URLSearchParams();
+		query.set('offset', String(offset));
 		if (searchStr) {
-			parts.push(`search=${encodeURIComponent(searchStr)}`);
+			query.set('search', encodeURIComponent(searchStr));
 		}
 		for (const [key, value] of this.filterByValues) {
-			parts.push(`filter-${key}=${encodeURIComponent(value)}`);
+			if (typeof value === 'string') {
+				query.set(`filter-${key}`, encodeURIComponent(value));
+			} else {
+				value.forEach(val => query.append(`filter-${key}`, encodeURIComponent(val)));
+			}
 		}
-		return parts.join('&');
+		return query;
 	}
 
-	protected async loadOptions(query: string, successCallback: Function) {
+	protected async loadOptions(query: URLSearchParams, successCallback: Function) {
 		const headers = new Headers();
 		headers.append('Accept', 'application/json');
-		const url = `${this.endpoint}?field=${this.fieldName}&${query}`;
-		const response = await fetch(url, {
+		query.set('field', this.fieldName!);
+		const response = await fetch(`${this.endpoint}?${query.toString()}`, {
 			method: 'GET',
 			headers: headers,
 		});
@@ -81,7 +95,7 @@ export abstract class IncompleteSelect {
 			}
 			successCallback(data.options);
 		} else {
-			console.error(`Failed to fetch from ${url} (status=${response.status})`);
+			console.error(`Failed to fetch from ${this.endpoint} (status=${response.status})`);
 		}
 	}
 }